[Fixes #4713] Attributes are not added in layers_attributes table when a layer is uploaded
 - Add test case

 - Add test case
