+++ geonode/geoserver/helpers.py
@@ -83,7 +83,7 @@ def check_geoserver_is_up():
        this is needed to be able to upload.
     """
     url = "%s" % ogc_server_settings.LOCATION
-    req, content = http_client.get(url)
+    req, content = http_client.get(url, user=_user)
     msg = ('Cannot connect to the GeoServer at %s\nPlease make sure you '
            'have started it.' % url)
     logger.debug(req)
@@ -537,6 +537,7 @@ def gs_slurp(
 
     if verbosity > 0:
         print >> console, "Inspecting the available layers in GeoServer ..."
+
     cat = Catalog(ogc_server_settings.internal_rest, _user, _password)
     if workspace is not None:
         workspace = cat.get_workspace(workspace)
@@ -894,7 +895,7 @@ def set_attributes_from_geoserver(layer, overwrite=False):
         dft_url = server_url + ("%s?f=json" % layer.alternate)
         try:
             # The code below will fail if http_client cannot be imported
-            req, body = http_client.get(dft_url)
+            req, body = http_client.get(dft_url, user=_user)
             body = json.loads(body)
             attribute_map = [[n["name"], _esri_types[n["type"]]]
                              for n in body["fields"] if n.get("name") and n.get("type")]
@@ -911,9 +912,8 @@ def set_attributes_from_geoserver(layer, overwrite=False):
                                                                   "typename": layer.alternate.encode('utf-8'),
                                                                   })
         try:
-            # The code below will fail if http_client cannot be imported  or
-            # WFS not supported
-            req, body = http_client.get(dft_url)
+            # The code below will fail if http_client cannot be imported or WFS not supported
+            req, body = http_client.get(dft_url, user=_user)
             doc = dlxml.fromstring(body)
             path = ".//{xsd}extension/{xsd}sequence/{xsd}element".format(
                 xsd="{http://www.w3.org/2001/XMLSchema}")
@@ -940,7 +940,7 @@ def set_attributes_from_geoserver(layer, overwrite=False):
                 "y": 1
             })
             try:
-                req, body = http_client.get(dft_url)
+                req, body = http_client.get(dft_url, user=_user)
                 soup = BeautifulSoup(body)
                 for field in soup.findAll('th'):
                     if(field.string is None):
@@ -961,7 +961,7 @@ def set_attributes_from_geoserver(layer, overwrite=False):
             "identifiers": layer.alternate.encode('utf-8')
         })
         try:
-            req, body = http_client.get(dc_url)
+            req, body = http_client.get(dc_url, user=_user)
             doc = dlxml.fromstring(body)
             path = ".//{wcs}Axis/{wcs}AvailableKeys/{wcs}Key".format(
                 wcs="{http://www.opengis.net/wcs/1.1.1}")
@@ -1543,7 +1543,7 @@ class OGC_Servers_Handler(object):
 def get_wms():
     wms_url = ogc_server_settings.internal_ows + \
         "?service=WMS&request=GetCapabilities&version=1.1.0"
-    req, body = http_client.get(wms_url)
+    req, body = http_client.get(wms_url, user=_user)
     _wms = WebMapService(wms_url, xml=body)
     return _wms
 
@@ -1574,7 +1574,7 @@ def wps_execute_layer_attribute_statistics(layer_name, field):
         method='POST',
         data=request,
         headers=headers,
-        user=ogc_server_settings.credentials.username)
+        user=_user)
 
     exml = dlxml.fromstring(content)
 
@@ -1608,7 +1608,7 @@ def _stylefilterparams_geowebcache_layer(layer_name):
     req, content = http_client.get(
         url,
         headers=headers,
-        user=ogc_server_settings.credentials.username)
+        user=_user)
     if req.status_code != 200:
         line = "Error {0} reading Style Filter Params GeoWebCache at {1}".format(
             req.status_code, url
@@ -1633,7 +1633,7 @@ def _stylefilterparams_geowebcache_layer(layer_name):
             url,
             data=body,
             headers=headers,
-            user=ogc_server_settings.credentials.username)
+            user=_user)
         if req.status_code != 200:
             line = "Error {0} writing Style Filter Params GeoWebCache at {1}".format(
                 req.status_code, url
@@ -1655,7 +1655,7 @@ def _invalidate_geowebcache_layer(layer_name, url=None):
         url,
         data=body,
         headers=headers,
-        user=ogc_server_settings.credentials.username)
+        user=_user)
 
     if req.status_code != 200:
         line = "Error {0} invalidating GeoWebCache at {1}".format(
